# metal : enable mat-vec kernels for bs <= 4